+++ examples/full-screen/simple-demos/vertical-split.py
@@ -3,33 +3,36 @@
 Vertical split example.
 """
 from quo.console import Console
+from quo.keys import Bind
 from quo.layout import FormattedTextControl, Window, VSplit, Layout
-from quo.keys import KeyBinder
+from quo.text import Text
 
 # 1. The layout
-left_text = "\nVertical-split example. Press 'q' to quit.\n\n(left pane.)"
-right_text = "\n(right pane.)"
+left_text = Text("""<b>Vertical-split  Press <red>'q'</red> to quit.
+
+(left pane.)</b>""")
+right_text = Text("""
+
+        <b>(right pane)</b>""")
 
 
 body = VSplit(
     [
         Window(FormattedTextControl(left_text)),
-        Window(width=1, char="|"),  # Vertical line in the middle.
+        Window(width=1, char="|", style="fg:green"),  # Vertical line in the middle.
         Window(FormattedTextControl(right_text)),
     ]
 )
 
 
 # 2. Key bindings
-kb = KeyBinder()
+bind = Bind()
 
-
-@kb.add("q")
+@bind.add("q")
 def _(event):
     "Quit application."
     event.app.exit()
 
 
 # 3. The `Application`
-Console(Layout(body), bind=kb, full_screen=True).run()
-
+Console(Layout(body), bind=bind, full_screen=True).run()